+++ app/src/main/kotlin/com/yapt/planttracker/domain/model/CareType.kt
@@ -7,5 +7,6 @@ enum class CareType {
     MIST,
     REPOT,
     NOTE,
-    PHOTO
+    PHOTO,
+    CUSTOM
 }